Lk/psy5 pnm update (#192)
* make thing compatible Matt's PNM ybus-reductions branch

* formatting

+++ test/test_utils/legacy_pf.jl
@@ -96,8 +96,8 @@ end
 # this function is for testing purposes only
 function _legacy_dSbus_dV(
     V::Vector{Complex{Float64}},
-    Ybus::SparseMatrixCSC{Complex{Float64}, Int64},
-)::Tuple{SparseMatrixCSC{Complex{Float64}, Int32}, SparseMatrixCSC{Complex{Float64}, Int32}}
+    Ybus::SparseMatrixCSC{Complex{Float32}, Int64},
+)::Tuple{SparseMatrixCSC{Complex{Float32}, Int32}, SparseMatrixCSC{Complex{Float32}, Int32}}
     diagV = SparseArrays.spdiagm(0 => V)
     diagVnorm = SparseArrays.spdiagm(0 => V ./ abs.(V))
     diagIbus = SparseArrays.spdiagm(0 => Ybus * V)
@@ -108,8 +108,8 @@ end
 
 # this function is for testing purposes only
 function _legacy_J(
-    dSbus_dVa::SparseMatrixCSC{Complex{Float64}, Int32},
-    dSbus_dVm::SparseMatrixCSC{Complex{Float64}, Int32},
+    dSbus_dVa::SparseMatrixCSC{Complex{Float32}, Int32},
+    dSbus_dVm::SparseMatrixCSC{Complex{Float32}, Int32},
     pvpq::Vector{Int64},
     pq::Vector{Int64},
 )
